Migrate to rebar3
This commit enables using cberl as a dependency from a rebar3 project.

The corrolary is that it might prevent using it from a rebar2 project. No attempt has been made to
keep the rebar2 usecase functional, but it should still work at this stage, as no
backward-incompatible changes to rabar.config have been made.

Rebar3 simplifies cberl's makefile a bit, brings some goodies, and seems to be the community's most
popular build system at this stage.